% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation_test.R
\name{autocorrelation_test}
\alias{autocorrelation_test}
\title{A function to the autocorrelation of amplific curve data 
from a quantitative PCR experiment}
\usage{
autocorrelation_test(y, n = 3, sig.level = 0.01)
}
\arguments{
\item{y}{is the cycle dependent fluorescence amplitude (y-axis).}

\item{n}{is}

\item{sig.level}{is the significance level for the correlation test.}
}
\description{
\code{autocorrelation_test} is a function an autocorrelation analysis
from a quantitative PCR experiment.
}
\examples{

# calculate head to tail ratio on noise (negative) amplification curve data

autocorrelation_test(y=rnorm(35))

}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{autocorrelation}
