% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plotter.R
\name{class_plotter}
\alias{class_plotter}
\title{A helper function to plot the frequency of classes of amplification curves}
\usage{
class_plotter(data, classes)
}
\arguments{
\item{data}{is the data set containing the amplification curves.}

\item{classes}{is the data set containing the classes of the amplification curves.}
}
\description{
\code{class_plotter} is a function to to plot the frequency of classes of 
amplification curves.
}
\examples{

# Use class_plotter to visualize the classified amplification curve data
library(qpcR)
library(data.table)
filename <- system.file("decission_res_htPCR.csv", package = "PCRedux")
class_plotter(data=htPCR, classes=as.data.frame(fread(filename)))

}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{classes}
\keyword{plot}
