% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlyreg.R
\name{earlyreg}
\alias{earlyreg}
\title{A function to calculate the slope and intercept of an amplific curve data 
from a quantitative PCR experiment.}
\usage{
earlyreg(x, y, range = 10)
}
\arguments{
\item{y}{is the cycle dependent fluorescence amplitude (y-axis).}
}
\description{
\code{earlyreg} is a function to calculate the slope and intercept of an 
amplific curve data from a quantitative PCR experiment.
}
\examples{

# Calculate slope and intercept on noise (negative) amplification curve data
# for the cycles 2 to 10.

earlyreg(1:35, y=rnorm(35), range=3:11)

}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{intercept}
\keyword{slope}
