% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head2tailratio.R
\name{head2tailratio}
\alias{head2tailratio}
\title{A function to calculate to head to tail ratio of amplification curve data 
from a quantitative PCR experiment}
\usage{
head2tailratio(y)
}
\arguments{
\item{y}{is the cycle dependent fluorescence amplitude (y-axis).}
}
\description{
\code{head2tailratio} is a function to calculate the ratio of the head and the
tail of a quantitative PCR amplification curve.
}
\examples{

# calculate head to tail ratio on noise (negative) amplification curve data

head2tailratio(y=rnorm(35))

}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{head}
\keyword{ratio}
\keyword{tail}
