% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hookreg.R
\name{hookreg}
\alias{hookreg}
\title{A function to calculate the slope and intercept of an amplification curve data 
from a quantitative PCR experiment at the end of the data stream.}
\usage{
hookreg(x, y)
}
\arguments{
\item{x}{is the cycle numbers (x-axis).}

\item{y}{is the cycle dependent fluorescence amplitude (y-axis).}
}
\description{
\code{hookreg} is a function to calculate the slope and intercept of an 
amplification curve data from a quantitative PCR experiment. The idea is that
a strong negative slope at the end of an amplification curve is indicative for
a hook effect (see Barratt and Mackay 2002).
}
\examples{

# Calculate slope and intercept on noise (negative) amplification curve data
# for the last eight cycles.

hookreg(x=1:35, y=rnorm(35))

}
\references{
K. Barratt, J.F. Mackay, \emph{Improving Real-Time PCR Genotyping 
Assays by Asymmetric Amplification}, J. Clin. Microbiol. 40 (2002) 1571–1572. 
doi:10.1128/JCM.40.4.1571-1572.2002.
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{hook}
\keyword{intercept}
\keyword{slope}
