% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hookregNL.R
\name{hookregNL}
\alias{hookregNL}
\title{hookregNL}
\usage{
hookregNL(x, y, plot = FALSE, level = 0.99, simple = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{plot}{PARAM_DESCRIPTION, Default: FALSE}

\item{level}{PARAM_DESCRIPTION, Default: 0.99}

\item{simple}{is a logical parameter. If TRUE (default) only the slope, 
confidence interval and decisions are shown as output 
(\code{\link[base]{data.frame}}). If FALSE, a \code{\link[base]{list}} 
including the 6-parameter model is the output.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
\code{hookregNL} is just a nice thing
}
\details{
DETAILS
}
\examples{
# Analyze data from the boggy data set for potential hook effect like 
# curvature
library(qpcR)
# has hook
res <- hookregNL(boggy[, 1], boggy[, 12])
res

# has no hook
res <- hookregNL(reps[, 1], reps[, 2])
res
}
\seealso{
\code{\link[qpcR]{pcrfit}}
 \code{\link[stats]{confint}}
}
\author{
Andrej-Nikolai Spiess, Stefan Roediger, Michal Burdukiewcz
}
