% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanrater.R
\name{humanrater2}
\alias{humanrater2}
\title{Human Rater 2.0}
\usage{
humanrater2()
}
\value{
No return value, called for side effects
}
\description{
Launches graphical user interface for the manual annotation of
large amplification curve data sets, similarly to
the \code{\link[chipPCR]{humanrater}} function.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

