% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mblrr.R
\name{mblrr}
\alias{mblrr}
\title{A function perform the Median based Local Robust Regression (mblrr) 
from a quantitative PCR experiment}
\usage{
mblrr(x, y, sig.level = 0.01)
}
\arguments{
\item{x}{is the cycles.}

\item{y}{is the cycle dependent fluorescence amplitude (y-axis).}

\item{sig.level}{is the significance level for the correlation test.}
}
\description{
\code{mblrr} is a function to perform the Median based Local Robust Regression (mblrr) 
from a quantitative PCR experiment.
}
\examples{

# Perfrom an mblrr analysis on noise (negative) amplification data of
# with 35 cycles.

mblrr(x=1:35, y=rnorm(35))

}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{regression}
\keyword{segmented}
