% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRedux.R
\docType{package}
\name{PCRedux-package}
\alias{PCRedux-package}
\alias{PCRedux}
\title{PCRedux - quantitative PCR machine learning helper tool}
\description{
\code{PCRedux} package is a toolbox for the analysis of qPCR data.
}
\section{ML}{

T.B.D.
}

\examples{
# use data set from package
plot(rnorm(35))
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
