\name{PCRedux_datasets}
\alias{RAS002_RAS003_amp}
\title{The (published) datasets implemented in PCRedux}

\description{
A compilation of datasets for method evaluation/comparison.
}

\usage{
RAS002_RAS003_amp
}

\details{
\bold{RAS002_RAS003_amp:}\cr
Setup: qPCR with replicates.\cr
Annotation: T.B.D.\cr
Hardware: CFX96.\cr
Details:\cr
RAS002_RAS003_amp: Primers, Taqman probes.\cr
}

\author{
Package authors
}

\references{
\bold{RAS002_RAS003_amp:}\cr
T.B.D.
}
\keyword{data}
