% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRedux_gui.R
\name{PCRedux_gui}
\alias{PCRedux_gui}
\title{PCRedux Graphical User Interface}
\usage{
PCRedux_gui()
}
\value{
No return value, called for side effects
}
\description{
Launches graphical user interface that generates numerous
features of a large amplification curve data set, similarly to
the \code{\link[PCRedux]{encu}} function.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

