% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrfit_parallel.R
\name{pcrfit_parallel}
\alias{pcrfit_parallel}
\title{A function to calculate numerous features from amplification curve data 
from a quantitative PCR experiment.}
\usage{
pcrfit_parallel(data, less_cores = 0, detection_chemistry = NA,
  device = NA)
}
\arguments{
\item{data}{is the data set containing the cycles and fluorescence amplitudes.}

\item{less_cores}{defines the numbers of cores that should be left unused by this function. By default is pcrfit_parallel using all cores.}

\item{detection_chemistry}{contains additional meta information about the detection chemistry (e.g., probes, intercalating dye) that was used.}

\item{device}{contains additional meta information about the qPCR system that was used.}
}
\description{
\code{pcrfit_parallel} is a function to calculate numerous features of a large 
amplification curve data set. pcrfit_parallel makes use of parallelized code
to make use of multi-core architectures.
}
\examples{

# Calculate curve features of an amplification curve data set by using all 
# available CPU cores.

# T.B.D.

}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{intercept}
\keyword{slope}
