% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrfit_helper.R
\name{prcfit_single}
\alias{prcfit_single}
\title{FUNCTION_TITLE}
\usage{
prcfit_single(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
# Load the chipPCR package and analyse from the C126EG685 the first qPCR run
# "A01" (column 2).
library(chipPCR)
res <- prcfit_single(C126EG685[, 2])
}
\seealso{
\code{\link[bcp]{bcp}}
 \code{\link[chipPCR]{bg.max}},\code{\link[chipPCR]{amptester}},\code{\link[chipPCR]{smoother}}
 \code{\link[ecp]{e.agglo}}
 \code{\link[MBmca]{diffQ}},\code{\link[MBmca]{mcaPeaks}},\code{\link[MBmca]{diffQ2}}
 \code{\link{head2tailratio}},\code{\link{earlyreg}},\code{\link{hookreg}},\code{\link{mblrr}},\code{\link{autocorrelation_test}}
 \code{\link[pracma]{polyarea}}
 \code{\link[qpcR]{pcrfit}},\code{\link[qpcR]{takeoff}},\code{\link[qpcR]{LRE}},\code{\link[qpcR]{sliwin}},\code{\link[qpcR]{efficiency}}
 \code{\link[base]{diff}}
 \code{\link[stats]{quantile}}
}
