% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR2fdata.R
\name{qPCR2fdata}
\alias{qPCR2fdata}
\title{A helper function to convert amplification curve data to the fdata format.}
\usage{
qPCR2fdata(data)
}
\arguments{
\item{data}{is a data set containing the amplification cycles (1. column) 
and the fluorescence (subsequent columns).}
}
\description{
\code{qPCR2fdata} is a helper function
}
\examples{

# Calculate slope and intercept on noise (negative) amplification curve data
# for the last eight cycles.
library(qpcR)
library(fda.usc)
library(magrittr)

res_fdata <- qPCR2fdata(testdat)
res_fdata_colnames <- testdat[-1] \%>\% colnames()
data_colors <- rainbow(length(res_fdata_colnames), alpha=0.5)

par(mfrow=c(1,2))
res_fdata \%>\% plot(., xlab="cycles", ylab="RFU", main="testdat", type="l", 
                   lty=1, lwd=2, col=data_colors)
legend("topleft", as.character(res_fdata_colnames), pch=19, 
         col=data_colors, bty="n", ncol=2)
res_fdata_hclust <- metric.hausdorff(res_fdata)
plot(hclust(as.dist(res_fdata_hclust))) 
 
}
\references{
M. Febrero-Bande, M.O. de la Fuente, others, \emph{Statistical 
computing in functional data analysis: The R package fda.usc}, Journal of 
Statistical Software. 51 (2012) 1–28. http://www.jstatsoft.org/v51/i04/
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{fdata}
