% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PCRedux.R
\name{run_PCRedux}
\alias{run_PCRedux}
\title{PCRedux app}
\usage{
run_PCRedux()
}
\value{
null.
}
\description{
A graphical user interface for compute the properties of amplification curves.
}
\note{
Any ad-blocking software may cause malfunctions.
}
\seealso{
\code{\link{encu}}, \code{\link[shiny]{runApp}}.
}
