% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visdat_pcrfit_parallel.R
\name{visdat_pcrfit_parallel}
\alias{visdat_pcrfit_parallel}
\title{Visualises the content of data from an analysis with the pcrfit_parallel function}
\usage{
visdat_pcrfit_parallel(data, type = "all", interactive = TRUE)
}
\arguments{
\item{data}{contains the result from an analysis with the \link[PCRedux]{pcrfit_parallel} function}

\item{type}{specifies of all or only a subset of results should be shown, 
Default: 'all'. Alternatives 'qpcR' or 'amptester'}

\item{interactive}{is a logical parameter, which indicates if the plot 
should be interactive, Default: TRUE}
}
\description{
gives a `ggplot` or interactive `ggplotly` object (default) of a 'pcrfit_parallel' data frame. The function is based on the \link[visdat]{vis_dat} function by Tierney (2017). The colored indicators show the class (e.g., missing values, factors).
}
\details{
type 'all' shows all results from the analysis by the 
\link[PCRedux]{pcrfit_parallel} function.
}
\examples{

# Calculate curve features of an amplification curve data. Note that not all 
# available CPU cores are used. If need set "all" to use all available cores.
library(qpcR)
# take the samples F1.1 (positive), F1.3 (negative), F3.1 (positive), 
# F3.3 (negative), F5.4 (positive) and F6.2 (negative) for this example.

test_data <- testdat[, c(1,2,4,10,12,21,23)]

# Plot the amplification curves

matplot(test_data[, 1], test_data[, -1], xlab="Cycle", ylab="RFU", 
        main="testdat dataset", type="l", lty=1, lwd=2, col=1:6)
legend("topleft", paste(colnames(test_data)[-1], rep(c("pos", "neg"), 3)), 
       pch=19, col=1:6)

# Analyze the amplification curves with the pcrfit_parallel function
res <- pcrfit_parallel(test_data)
res

# Show all results in an interactive plot
visdat_pcrfit_parallel(res, type="all")
}
\references{
N. Tierney, visdat: Visualising Whole Data Frames, \emph{The 
Journal of Open Source Software}. 2 (2017). doi:10.21105/joss.00355.
}
\seealso{
\code{\link[plotly]{ggplotly}}
 \code{\link[visdat]{vis_dat}}
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
