% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decission_modus.R
\name{decission_modus}
\alias{decission_modus}
\title{A function to get a decission (modus) from a vector of classes}
\usage{
decission_modus(data, variables = c("a", "n", "y"), max_freq = TRUE)
}
\arguments{
\item{data}{is a table contining the classes.}

\item{variables}{is the class to look for.}

\item{max_freq}{delivers either the most ocurring class or a summary.}
}
\description{
\code{decission_modus} is a function that can be used to find the most frequent
(modus) decission. The classes can be defined by the user (e.g., a", "n", "y" 
-> "ambiguos", "negaive", "positive"). This function is usefuel if large 
collections of varying decission (e.g., "a", "a", "a", "n", "n") need to be 
condensed to a single decission (3 x "a", 2 x "n" -> "a").
}
\examples{

# #'
library(data.table)
library(magrittr)
filename <- system.file("decission_res_testdat.csv", package = "PCRedux")
my_data <- as.data.frame(fread(filename))
head(my_data)

dec <- lapply(1L:nrow(my_data), function(i) {
       decission_modus(my_data[i, 2:4])
}) \%>\% unlist

names(dec) <- my_data[, 1]
dec
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{decission}
\keyword{modus}
