% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRedux.R
\docType{package}
\name{PCRedux-package}
\alias{PCRedux-package}
\alias{PCRedux}
\title{PCRedux - quantitative PCR machine learning helper tool}
\description{
\code{PCRedux} package is a toolbox for the analysis of sigmoid curve (qPCR) data.
}
\section{ML}{

T.B.D.
}

\examples{
# Use the mblrr functopm to analyse 
mblrr(x=1:35, y=rnorm(35))
}
\author{
Stefan Roediger, Michal Burdukiewcz, Andrej-Nikolai Spiess
}
