% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRedux.R
\docType{package}
\name{PCRedux-package}
\alias{PCRedux-package}
\alias{l4}
\alias{PCRedux}
\title{PCRedux - quantitative PCR Data Mining and Machine Learning Helper Tool}
\format{An object of class \code{list} of length 11.}
\usage{
l4
}
\description{
\code{PCRedux} package is a toolbox for the analysis of sigmoid curve (qPCR) data.
}
\section{Machine learning}{

In machine learning and statistics, the classification should be used to identify a new unknown observation. This observation is assigned to a number of categories. One basis is training data sets containing observations with known classes. Using the example of sigmoid amplification curves, this could be an assignment to the class "negative","ambiguous" or "positive". Basically, a number of descriptors (e. g., characteristics of curvature) are required to be able to assign classes. This package contains functions for extracting characteristics. In addition, the package contains data sets of classified amplification curves.
}

\examples{
# Use the mblrr function to analyse amplification curves
library(qpcR)
mblrr(x=boggy[, 1], y=boggy[, 2])
}
\author{
Stefan Roediger, Michal Burdukiewcz, Andrej-Nikolai Spiess, Konstantin A. Blagodatskikh
}
\keyword{datasets}
