% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrfit_parallel.R
\name{pcrfit_parallel}
\alias{pcrfit_parallel}
\title{A function to calculate numerous features from amplification curve data 
from a quantitative PCR experiment.}
\usage{
pcrfit_parallel(data, n_cores = 1, detection_chemistry = NA, device = NA)
}
\arguments{
\item{data}{is the data set containing the cycles and fluorescence amplitudes.}

\item{n_cores}{defines the numbers of cores that should be left unused by this 
function. By default pcrfit_parallel is using only two cores. Use
\code{"all"} to use all available cores.}

\item{detection_chemistry}{contains additional meta information about the 
detection chemistry (e.g., probes, intercalating dye) that was used.}

\item{device}{contains additional meta information about the qPCR system that 
was used.}
}
\value{
The output of the pcrfit_parallel function is 
identical to the \code{\link[PCRedux]{pcrfit_single}} function.
}
\description{
pcrfit_parallel is a function to calculate numerous 
features of a large amplification curve data set. 
pcrfit_parallel makes use of parallelized code to make 
use of multi-core architectures. The \code{\link[PCRedux]{pcrfit_single}} is 
performing the analysis for a single process. pcrfit_parallel
}
\examples{

# Calculate curve features of an amplification curve data. Note that not all 
# available CPU cores are used. If need set "all" to use all available cores.
# In this example the testdat dataset from the qpcR package is used.
# The samples F1.1 and F1.2 are positive amplification curves. The samples 
# F1.3 and F1.4 are negative.

library(qpcR)
res_pcrfit_parallel <- pcrfit_parallel(testdat[, 1:5])
res_pcrfit_parallel
 
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
\keyword{intercept}
\keyword{normalization}
\keyword{preprocessing}
\keyword{slope}
