% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrfit_single.R
\name{prcfit_single}
\alias{prcfit_single}
\title{prcfit_single - A function to extract features from an amplification curve}
\usage{
prcfit_single(x)
}
\arguments{
\item{x}{is the data set containing the fluorescence amplitudes.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
The prcfit_single is responsible for the 
extraction of features from amplification curve data. This function is the basis
of the \code{\link[PCRedux]{pcrfit_parallel}} function. The later performs the 
parallelized analysis of amplification curve data
}
\details{
DETAILS
}
\examples{
# Load the chipPCR package and analyse from the C126EG685 the first qPCR run
# "A01" (column 2).
library(chipPCR)
res <- prcfit_single(C126EG685[, 2])
}
\seealso{
\code{\link[bcp]{bcp}}
 \code{\link[chipPCR]{bg.max}},\code{\link[chipPCR]{amptester}},\code{\link[chipPCR]{smoother}}
 \code{\link[ecp]{e.agglo}}
 \code{\link[MBmca]{diffQ}},\code{\link[MBmca]{mcaPeaks}},\code{\link[MBmca]{diffQ2}}
 \code{\link{head2tailratio}},\code{\link{earlyreg}},\code{\link{hookreg}},\code{\link{mblrr}},\code{\link{autocorrelation_test}}
 \code{\link[pracma]{polyarea}}
 \code{\link[qpcR]{pcrfit}},\code{\link[qpcR]{takeoff}},\code{\link[qpcR]{LRE}},\code{\link[qpcR]{sliwin}},\code{\link[qpcR]{efficiency}}
 \code{\link[base]{diff}}
 \code{\link[stats]{quantile}}
}
