% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_PCRedux.R
\name{run_PCRedux}
\alias{run_PCRedux}
\title{PCRedux app}
\usage{
run_PCRedux()
}
\value{
null.

No return value, called for side effects
}
\description{
A graphical user interface for computing the properties of amplification curves. 
Take a look at the vignette to learn more about the different ways to start the app.
}
\note{
Any ad-blocking software may cause malfunctions.
}
\seealso{
\code{\link{encu}}, \code{\link[shiny]{runApp}}.
}
