% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visdat_pcrfit_parallel.R
\name{visdat_pcrfit_parallel}
\alias{visdat_pcrfit_parallel}
\title{Visualises the content of data from an analysis with the pcrfit_parallel function}
\usage{
visdat_pcrfit_parallel(data, type = "all", interactive = TRUE)
}
\arguments{
\item{data}{contains the result from an analysis with the \link[PCRedux]{pcrfit_parallel} function}

\item{type}{specifies of all or only a subset of results should be shown, 
Default: 'all'. Alternatives 'qpcR' or 'amptester'}

\item{interactive}{is a logical parameter, which indicates if the plot 
should be interactive, Default: TRUE}
}
\description{
gives a `ggplot` or interactive `ggplotly` object (default) of a 'pcrfit_parallel' data frame. The function is based on the \link[visdat]{vis_dat} function by Tierney (2017). The colored indicators show the class (e.g., missing values, factors).
}
\details{
type 'all' shows all results from the analysis by the 
\link[PCRedux]{pcrfit_parallel} function.
}
\examples{

# Calculate curve features of an amplification curve data. Note that not all 
# available CPU cores are used. If need set "all" to use all available cores.
library(qpcR)
res_pcrfit_parallel <- pcrfit_parallel(boggy[, 1:7])
res_pcrfit_parallel

# Show all results in an interactive plot
visdat_pcrfit_parallel(res_pcrfit_parallel, type="all")
}
\references{
N. Tierney, visdat: Visualising Whole Data Frames, \emph{The 
Journal of Open Source Software}. 2 (2017). doi:10.21105/joss.00355.
}
\seealso{
\code{\link[plotly]{ggplotly}}
 \code{\link[visdat]{vis_dat}}
}
\author{
Stefan Roediger, Michal Burdukiewcz
}
